FIX include database to example controller
